InfoMundi/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pymysql
pandas
apscheduler
prefect
python-dotenv
